+++ notefile/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "2.20230212.0"
+__version__ = "2.20230215.0"
 __author__ = "Justin Winokur"
 
 import sys, os